requirements.txt
Pillow
numpy
click
rasterio[s3]>=1.0a11
Wand
